fix: resolve component pod vars from desired instanceset
