+++ pkg/controller/factory/builder_test.go
@@ -27,7 +27,6 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -102,14 +101,14 @@ var _ = Describe("builder", func() {
 		It("builds InstanceSet correctly", func() {
 			compDef, cluster, synthesizedComponent := newClusterObjs(nil)
 
-			its, err := BuildInstanceSet(synthesizedComponent, nil)
+			its, err := component.BuildInstanceSet(synthesizedComponent, nil)
 			Expect(err).Should(BeNil())
 			Expect(its).ShouldNot(BeNil())
 
 			By("set replicas = 0")
 			newComponent := *synthesizedComponent
 			newComponent.Replicas = 0
-			its, err = BuildInstanceSet(&newComponent, nil)
+			its, err = component.BuildInstanceSet(&newComponent, nil)
 			Expect(err).Should(BeNil())
 			Expect(its).ShouldNot(BeNil())
 			Expect(*its.Spec.Replicas).Should(Equal(int32(0)))
@@ -117,7 +116,7 @@ var _ = Describe("builder", func() {
 			By("set replicas = 2")
 			cluster.Spec.ComponentSpecs[0].Replicas = 2
 			synthesizedComp := newAllFieldsSynthesizedComponent(compDef, cluster)
-			its, err = BuildInstanceSet(synthesizedComp, nil)
+			its, err = component.BuildInstanceSet(synthesizedComp, nil)
 			Expect(err).Should(BeNil())
 			Expect(its).ShouldNot(BeNil())
 			Expect(*its.Spec.Replicas).Should(BeEquivalentTo(2))
@@ -150,7 +149,7 @@ var _ = Describe("builder", func() {
 				},
 			}
 
-			its, err := BuildInstanceSet(synthesizedComp, nil)
+			its, err := component.BuildInstanceSet(synthesizedComp, nil)
 
 			Expect(err).Should(Succeed())
 			Expect(its.Labels).Should(HaveKeyWithValue(constant.KBAppShardingNameLabelKey, shardingName))
@@ -225,111 +224,4 @@ var _ = Describe("builder", func() {
 		})
 	})
 
-	Context("cluster default resources", func() {
-		AfterEach(func() {
-			viper.Set(constant.CfgKeyClusterDefaultResources, "")
-		})
-
-		newInstanceSet := func() *workloads.InstanceSet {
-			return &workloads.InstanceSet{
-				Spec: workloads.InstanceSetSpec{
-					Template: corev1.PodTemplateSpec{
-						Spec: corev1.PodSpec{
-							Containers: []corev1.Container{
-								{Name: "main"},
-								{Name: "sidecar"},
-							},
-							InitContainers: []corev1.Container{
-								{Name: "init"},
-							},
-						},
-					},
-				},
-			}
-		}
-
-		It("does not inject zero resources into init and sidecar containers when not configured", func() {
-			its := newInstanceSet()
-
-			Expect(setDefaultResourceLimits(its)).Should(Succeed())
-
-			Expect(its.Spec.Template.Spec.Containers[0].Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("0")))
-			Expect(its.Spec.Template.Spec.Containers[1].Resources.Requests).Should(BeNil())
-			Expect(its.Spec.Template.Spec.Containers[1].Resources.Limits).Should(BeNil())
-			Expect(its.Spec.Template.Spec.InitContainers[0].Resources.Requests).Should(BeNil())
-			Expect(its.Spec.Template.Spec.InitContainers[0].Resources.Limits).Should(BeNil())
-		})
-
-		It("keeps zero resource limit behavior when zero is true", func() {
-			viper.Set(constant.CfgKeyClusterDefaultResources, `{"zero":true,"requests":{},"limits":{}}`)
-			its := newInstanceSet()
-
-			Expect(setDefaultResourceLimits(its)).Should(Succeed())
-
-			Expect(its.Spec.Template.Spec.Containers[0].Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("0")))
-			Expect(its.Spec.Template.Spec.Containers[1].Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("0")))
-			Expect(its.Spec.Template.Spec.InitContainers[0].Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("0")))
-		})
-
-		It("applies configured resources to init and sidecar containers", func() {
-			viper.Set(constant.CfgKeyClusterDefaultResources, `{"zero":true,"requests":{"cpu":"10m","memory":"16Mi"},"limits":{"cpu":"100m","memory":"64Mi"}}`)
-			its := newInstanceSet()
-
-			Expect(setDefaultResourceLimits(its)).Should(Succeed())
-
-			main := its.Spec.Template.Spec.Containers[0]
-			sidecar := its.Spec.Template.Spec.Containers[1]
-			initContainer := its.Spec.Template.Spec.InitContainers[0]
-			Expect(main.Resources.Requests).Should(BeNil())
-			Expect(main.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("0")))
-			Expect(sidecar.Resources.Requests).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("10m")))
-			Expect(sidecar.Resources.Requests).Should(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("16Mi")))
-			Expect(sidecar.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("100m")))
-			Expect(sidecar.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("64Mi")))
-			Expect(initContainer.Resources.Requests).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("10m")))
-			Expect(initContainer.Resources.Requests).Should(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("16Mi")))
-			Expect(initContainer.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("100m")))
-			Expect(initContainer.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("64Mi")))
-		})
-
-		It("lets configured resource names override zero by resource name", func() {
-			viper.Set(constant.CfgKeyClusterDefaultResources, `{"zero":true,"requests":{"cpu":"10m"},"limits":{}}`)
-			its := newInstanceSet()
-
-			Expect(setDefaultResourceLimits(its)).Should(Succeed())
-
-			sidecar := its.Spec.Template.Spec.Containers[1]
-			initContainer := its.Spec.Template.Spec.InitContainers[0]
-			Expect(sidecar.Resources.Requests).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("10m")))
-			Expect(sidecar.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("10m")))
-			Expect(sidecar.Resources.Requests).ShouldNot(HaveKey(corev1.ResourceMemory))
-			Expect(sidecar.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("0")))
-			Expect(initContainer.Resources.Requests).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("10m")))
-			Expect(initContainer.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("10m")))
-			Expect(initContainer.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("0")))
-		})
-
-		It("does not override sidecar resource values already set by definitions", func() {
-			viper.Set(constant.CfgKeyClusterDefaultResources, `{"zero":true,"requests":{"cpu":"10m","memory":"16Mi"},"limits":{"cpu":"100m","memory":"64Mi"}}`)
-			its := newInstanceSet()
-			its.Spec.Template.Spec.Containers[1].Resources.Requests = corev1.ResourceList{
-				corev1.ResourceCPU: resource.MustParse("250m"),
-			}
-
-			Expect(setDefaultResourceLimits(its)).Should(Succeed())
-
-			sidecar := its.Spec.Template.Spec.Containers[1]
-			Expect(sidecar.Resources.Requests).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("250m")))
-			Expect(sidecar.Resources.Requests).Should(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("16Mi")))
-			Expect(sidecar.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("250m")))
-			Expect(sidecar.Resources.Limits).Should(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("64Mi")))
-		})
-
-		It("returns an error when cluster default resources are invalid", func() {
-			viper.Set(constant.CfgKeyClusterDefaultResources, `{"zero":`)
-			its := newInstanceSet()
-
-			Expect(setDefaultResourceLimits(its)).ShouldNot(Succeed())
-		})
-	})
 })
